etl_requirements.txt
requests
python-dateutil
boto3
s3fs
pandas
apache-airflow-providers-cncf-kubernetes